Initial public commit
Archodex is a source-available project, and this is its first public
commit! But it was iteratively developed over time, and, unfortunately,
the original commit history has a few confidential values that we don't
want to expose publicly (e.g. private development resource IDs). This
commit is a squash commit of the latest code in the repository as of the
timestamp of the commit.