Improve GHA build workflow
This commit improves the GHA workflow by condensing the build down to a
single job that compiles the server for both x86_64 and aarch64
architectures simultaneously. This speeds up the overall build process
and simplifies container builds, pushes, and tagging.

This also moves from using the musl libc to standard glibc. The only
reasonably available musl cross toolchain is quite old (based on gcc v9)
and we can get by with glibc dynamic base chainguard container images.
The benefit to having a fully static binary is not worth the complexity
of musl. With glibc, we can also use mold for speedier linking in GHA,
too.