Refactor tracing for better logs
We now use `tracing::instrument(err)` in fallible functions to log any
time an error occurs along with relevant parameters and spans of
context. Further, when an unexpected error bubbles up to where we
convert it to an HTTP `PublicError`, we also print the full stack trace
(assuming the `RUST_BACKTRACE` env var is set to `1`).