Fix devcontainer perm issues for cargo git deps
We attempted to use the same volume for cargo git deps across this
project and archodex-agent, but because the agent devcontainer runs as
root and this container runs as the vscode user we end up with
permissions conflicts for git deps. This uses a separate volume for git
deps in the backend, and we'll create a namespaced volume for the agent,
too.