Implement self-hosted API private key support
In self-hosted mode we use either the API private key material from the
`ARCHODEX_API_PRIVATE_KEY` environment variable or from the account
database record. If neither exists we panic. If both exist we also
panic, as this is almost certainly a misconfiguration.

The purpose of the `ARCHODEX_API_PRIVATE_KEY` is to allow the key
material to be stored elsewhere outside of the database, but if it isn't
set then we generate key material when the account is created and save
it in the database.