Refactor transaction usage
We were inconsistently using transactions. We were never doing anything
problematic (e.g. not using a transaction when one is necessary), but we
were declaring transactions when we didn't need to be. There's no need
to declare a transaction when submitting a single query, whether or not
the query is read-only or write. This commit removes these unnecessary
transaction declarations, and refactors some of the query traits to be
callable directly on database connections.